third_party/blink/web_tests/platform/generic/fast/css/pseudo-default-003-expected.txt

This test performs a check for the :default CSS selector on multiple submit buttons, part 2.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Submit  
PASS document.defaultView.getComputedStyle(v[0], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(255, 0, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(255, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
